server/requirements.txt
flask==3.1.0
flask-cors
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
werkzeug==3.1.3
PyJWT==2.10.0
python-dotenv==1.0.1
pgvector==0.2.4